Implemented rss reader library.
Most rss and atom feed formats can now be read. As well as this, error
handling has been incorporated to ensure that if a feed cannot be read,
the user is made aware that an issue has occurred.
Ignored public cache files and updated exception blocker so that it
only reacts to E_ERROR and not warnings or notices.